Fixed therapy screen bug, reset broken player script, added in slower repair and attack speed (maybe, still WIP) based on low food and fuel levels. Attack damage increases by 0.5 everytime scene starts, got up to 30.5 when I was testing stuff.

+++ Sea of Stars/Assets/Scripts/GameManager.cs
@@ -9,8 +9,8 @@ using UnityEngine.SceneManagement;
 public class GameManager : MonoBehaviour
 {
     [Header("Resource Data")]
-    public int fuelCount = 10; // This is separate to prevent the header from appearing above all 3 attributes
-    public int foodCount = 10, luminosityCount = 10, minLuminosity = 0;
+    public float fuelCount = 10; // This is separate to prevent the header from appearing above all 3 attributes
+    public float foodCount = 10, luminosityCount = 10, minLuminosity = 0;
     public int crewCount = 0;
 
     private int foodCost;
